feat(anthropicadministrator): add modular architecture and missing AI features
