+++ packages/anthropicadministrator/endpoints/api-keys.ts
@@ -3,7 +3,7 @@ import { cacheEntity, cacheList, callAdminApi, compact } from './shared';
 import type {
 	ApiKey,
 	AnthropicAdministratorEndpointOutputs as Outputs,
-} from './types';
+} from './types/index';
 
 const BASE = '/v1/organizations/api_keys';
