+++ packages/anthropicadministrator/endpoints/invites.ts
@@ -9,7 +9,7 @@ import {
 import type {
 	Invite,
 	AnthropicAdministratorEndpointOutputs as Outputs,
-} from './types';
+} from './types/index';
 
 const BASE = '/v1/organizations/invites';
