+++ packages/anthropicadministrator/endpoints/workspace-members.ts
@@ -9,7 +9,7 @@ import {
 import type {
 	AnthropicAdministratorEndpointOutputs as Outputs,
 	WorkspaceMember,
-} from './types';
+} from './types/index';
 
 const BASE = '/v1/organizations/workspaces';
