+++ packages/anthropicadministrator/index.ts
@@ -1,5 +1,4 @@
 import type {
-	AuthTypes,
 	BindEndpoints,
 	CorsairEndpoint,
 	CorsairErrorHandler,
@@ -16,6 +15,8 @@ import { AuthMissingError } from 'corsair/core';
 import {
 	ApiKeysEndpoints,
 	InvitesEndpoints,
+	MessagesEndpoints,
+	ModelsEndpoints,
 	OrganizationEndpoints,
 	UsersEndpoints,
 	WorkspaceMembersEndpoints,
@@ -24,11 +25,11 @@ import {
 import type {
 	AnthropicAdministratorEndpointInputs,
 	AnthropicAdministratorEndpointOutputs,
-} from './endpoints/types';
+} from './endpoints/types/index';
 import {
 	AnthropicAdministratorEndpointInputSchemas,
 	AnthropicAdministratorEndpointOutputSchemas,
-} from './endpoints/types';
+} from './endpoints/types/index';
 import { errorHandlers } from './error-handlers';
 import { AnthropicAdministratorSchema } from './schema';
 
@@ -73,202 +74,16 @@ const anthropicAdministratorEndpointsNested = {
 	workspaces: WorkspacesEndpoints,
 	workspaceMembers: WorkspaceMembersEndpoints,
 	apiKeys: ApiKeysEndpoints,
+	messages: MessagesEndpoints,
+	models: ModelsEndpoints,
 } as const;
 
 const anthropicAdministratorWebhooksNested = {} as const;
 
-export const anthropicAdministratorEndpointSchemas = {
-	'organization.getOrganization': {
-		input: AnthropicAdministratorEndpointInputSchemas.getOrganization,
-		output: AnthropicAdministratorEndpointOutputSchemas.getOrganization,
-	},
-	'users.listUsers': {
-		input: AnthropicAdministratorEndpointInputSchemas.listUsers,
-		output: AnthropicAdministratorEndpointOutputSchemas.listUsers,
-	},
-	'users.getUser': {
-		input: AnthropicAdministratorEndpointInputSchemas.getUser,
-		output: AnthropicAdministratorEndpointOutputSchemas.getUser,
-	},
-	'users.updateUser': {
-		input: AnthropicAdministratorEndpointInputSchemas.updateUser,
-		output: AnthropicAdministratorEndpointOutputSchemas.updateUser,
-	},
-	'users.removeUser': {
-		input: AnthropicAdministratorEndpointInputSchemas.removeUser,
-		output: AnthropicAdministratorEndpointOutputSchemas.removeUser,
-	},
-	'invites.listInvites': {
-		input: AnthropicAdministratorEndpointInputSchemas.listInvites,
-		output: AnthropicAdministratorEndpointOutputSchemas.listInvites,
-	},
-	'invites.createInvite': {
-		input: AnthropicAdministratorEndpointInputSchemas.createInvite,
-		output: AnthropicAdministratorEndpointOutputSchemas.createInvite,
-	},
-	'invites.getInvite': {
-		input: AnthropicAdministratorEndpointInputSchemas.getInvite,
-		output: AnthropicAdministratorEndpointOutputSchemas.getInvite,
-	},
-	'invites.deleteInvite': {
-		input: AnthropicAdministratorEndpointInputSchemas.deleteInvite,
-		output: AnthropicAdministratorEndpointOutputSchemas.deleteInvite,
-	},
-	'workspaces.listWorkspaces': {
-		input: AnthropicAdministratorEndpointInputSchemas.listWorkspaces,
-		output: AnthropicAdministratorEndpointOutputSchemas.listWorkspaces,
-	},
-	'workspaces.createWorkspace': {
-		input: AnthropicAdministratorEndpointInputSchemas.createWorkspace,
-		output: AnthropicAdministratorEndpointOutputSchemas.createWorkspace,
-	},
-	'workspaces.getWorkspace': {
-		input: AnthropicAdministratorEndpointInputSchemas.getWorkspace,
-		output: AnthropicAdministratorEndpointOutputSchemas.getWorkspace,
-	},
-	'workspaces.updateWorkspace': {
-		input: AnthropicAdministratorEndpointInputSchemas.updateWorkspace,
-		output: AnthropicAdministratorEndpointOutputSchemas.updateWorkspace,
-	},
-	'workspaces.archiveWorkspace': {
-		input: AnthropicAdministratorEndpointInputSchemas.archiveWorkspace,
-		output: AnthropicAdministratorEndpointOutputSchemas.archiveWorkspace,
-	},
-	'workspaceMembers.listWorkspaceMembers': {
-		input: AnthropicAdministratorEndpointInputSchemas.listWorkspaceMembers,
-		output: AnthropicAdministratorEndpointOutputSchemas.listWorkspaceMembers,
-	},
-	'workspaceMembers.createWorkspaceMember': {
-		input: AnthropicAdministratorEndpointInputSchemas.createWorkspaceMember,
-		output: AnthropicAdministratorEndpointOutputSchemas.createWorkspaceMember,
-	},
-	'workspaceMembers.getWorkspaceMember': {
-		input: AnthropicAdministratorEndpointInputSchemas.getWorkspaceMember,
-		output: AnthropicAdministratorEndpointOutputSchemas.getWorkspaceMember,
-	},
-	'workspaceMembers.updateWorkspaceMember': {
-		input: AnthropicAdministratorEndpointInputSchemas.updateWorkspaceMember,
-		output: AnthropicAdministratorEndpointOutputSchemas.updateWorkspaceMember,
-	},
-	'workspaceMembers.deleteWorkspaceMember': {
-		input: AnthropicAdministratorEndpointInputSchemas.deleteWorkspaceMember,
-		output: AnthropicAdministratorEndpointOutputSchemas.deleteWorkspaceMember,
-	},
-	'apiKeys.listApiKeys': {
-		input: AnthropicAdministratorEndpointInputSchemas.listApiKeys,
-		output: AnthropicAdministratorEndpointOutputSchemas.listApiKeys,
-	},
-	'apiKeys.getApiKey': {
-		input: AnthropicAdministratorEndpointInputSchemas.getApiKey,
-		output: AnthropicAdministratorEndpointOutputSchemas.getApiKey,
-	},
-	'apiKeys.updateApiKey': {
-		input: AnthropicAdministratorEndpointInputSchemas.updateApiKey,
-		output: AnthropicAdministratorEndpointOutputSchemas.updateApiKey,
-	},
-} satisfies RequiredPluginEndpointSchemas<
-	typeof anthropicAdministratorEndpointsNested
->;
+import { anthropicAdministratorEndpointSchemas } from './meta';
+import { anthropicAdministratorEndpointMeta } from './meta-descriptions';
 
 const defaultAuthType = 'api_key' as const;
-
-const anthropicAdministratorEndpointMeta = {
-	'organization.getOrganization': {
-		riskLevel: 'read',
-		description: 'Get the organization associated with the Admin API key',
-	},
-	'users.listUsers': {
-		riskLevel: 'read',
-		description:
-			'List organization members, optionally filtered by email or role',
-	},
-	'users.getUser': {
-		riskLevel: 'read',
-		description: 'Get a single organization member by user ID',
-	},
-	'users.updateUser': {
-		riskLevel: 'write',
-		description: "Change an organization member's role",
-	},
-	'users.removeUser': {
-		riskLevel: 'destructive',
-		irreversible: true,
-		description: 'Remove a member from the organization',
-	},
-	'invites.listInvites': {
-		riskLevel: 'read',
-		description: 'List organization invites',
-	},
-	'invites.createInvite': {
-		riskLevel: 'write',
-		description: 'Invite someone to the organization with a given role',
-	},
-	'invites.getInvite': {
-		riskLevel: 'read',
-		description: 'Get a single invite by ID',
-	},
-	'invites.deleteInvite': {
-		riskLevel: 'destructive',
-		irreversible: true,
-		description: 'Delete a pending organization invite',
-	},
-	'workspaces.listWorkspaces': {
-		riskLevel: 'read',
-		description: 'List workspaces, optionally including archived ones',
-	},
-	'workspaces.createWorkspace': {
-		riskLevel: 'write',
-		description: 'Create a workspace',
-	},
-	'workspaces.getWorkspace': {
-		riskLevel: 'read',
-		description: 'Get a single workspace by ID',
-	},
-	'workspaces.updateWorkspace': {
-		riskLevel: 'write',
-		description: 'Update a workspace name, tags or data residency',
-	},
-	'workspaces.archiveWorkspace': {
-		riskLevel: 'destructive',
-		irreversible: true,
-		description: 'Archive a workspace',
-	},
-	'workspaceMembers.listWorkspaceMembers': {
-		riskLevel: 'read',
-		description: 'List members of a workspace',
-	},
-	'workspaceMembers.createWorkspaceMember': {
-		riskLevel: 'write',
-		description: 'Add an organization member to a workspace with a role',
-	},
-	'workspaceMembers.getWorkspaceMember': {
-		riskLevel: 'read',
-		description: 'Get a single workspace member',
-	},
-	'workspaceMembers.updateWorkspaceMember': {
-		riskLevel: 'write',
-		description: "Change a workspace member's role",
-	},
-	'workspaceMembers.deleteWorkspaceMember': {
-		riskLevel: 'destructive',
-		irreversible: true,
-		description: 'Remove a member from a workspace',
-	},
-	'apiKeys.listApiKeys': {
-		riskLevel: 'read',
-		description: 'List organization API keys, optionally filtered by status',
-	},
-	'apiKeys.getApiKey': {
-		riskLevel: 'read',
-		description: 'Get a single API key by ID',
-	},
-	'apiKeys.updateApiKey': {
-		riskLevel: 'write',
-		description: 'Rename an API key or change its active status',
-	},
-} as const satisfies RequiredPluginEndpointMeta<
-	typeof anthropicAdministratorEndpointsNested
->;
 
 export const anthropicAdministratorAuthConfig = {
 	api_key: {},
@@ -319,9 +134,15 @@ export function anthropicadministrator<
 		hooks: options.hooks,
 		webhookHooks: undefined,
 		endpoints: anthropicAdministratorEndpointsNested,
+		endpointMeta:
+			anthropicAdministratorEndpointMeta satisfies RequiredPluginEndpointMeta<
+				typeof anthropicAdministratorEndpointsNested
+			>,
 		webhooks: anthropicAdministratorWebhooksNested,
-		endpointMeta: anthropicAdministratorEndpointMeta,
-		endpointSchemas: anthropicAdministratorEndpointSchemas,
+		endpointSchemas:
+			anthropicAdministratorEndpointSchemas satisfies RequiredPluginEndpointSchemas<
+				typeof anthropicAdministratorEndpointsNested
+			>,
 		pluginWebhookMatcher: undefined,
 		errorHandlers: {
 			...errorHandlers,
@@ -369,8 +190,10 @@ export type {
 	AnthropicAdministratorEndpointOutputs,
 	ApiKey,
 	Invite,
+	Message,
+	Model,
 	Organization,
 	User,
 	Workspace,
 	WorkspaceMember,
-} from './endpoints/types';
+} from './endpoints/types/index';
